spd/requirements.txt
prometheus-client==0.15.0
PyYAML==6.0
